+++ packages/bundle/src/initialize.ts
@@ -15,6 +15,7 @@ import { getWidgets, bulkAddWidgets } from './core/widgets';
 import { renderWidgets } from './helpers/renderWidgets';
 import { observeDomNodes } from './helpers/observeDomNodes';
 import * as location from './core/location';
+import { injectComponents, injectStyles } from './utils/inject';
 
 /**
  * Create global namespace
@@ -38,9 +39,6 @@ const isReady = (() => {
   return true;
 })();
 
-const isString = (value) =>
-  typeof value === 'string' || value instanceof String;
-
 const _analyticsPromise = (() => {
   let resolve;
   const promise = new Promise((_resolve) => (resolve = _resolve));
@@ -62,23 +60,10 @@ export default async (_config, sentry) => {
     ...asyncConfig,
   };
 
-  // Register custom components
   if (cfg.components) {
-    const extra = Object.keys(cfg.components).reduce(
-      (acc, k) => ({
-        ...acc,
-        [k]: isString(cfg.components[k])
-          ? new Function('return ' + cfg.components[k])()
-          : cfg.components[k],
-      }),
-      {}
-    );
-    await __root.invalidate();
-    window.findifyJsonp.push([['extra'], extra]);
+    await injectComponents(cfg.components);
     delete cfg.components;
-    debug('bundle')('customizations:', extra);
   }
-
 
   __root.config = fromJS(cfg);
   __root.sentry = sentry;
@@ -162,36 +147,8 @@ export default async (_config, sentry) => {
     );
     window.addEventListener('message', async ({ data }) => {
       const { type, response } = data;
-      // On extension load and on component saved (onCommit), staging components are saved and sent.
-      if (type === 'components') {
-        console.log('[MJS]: inject updated components');
-        const content = Object.keys(response).reduce((acc, c) =>
-          [...acc, `"${response[c].hash}": ${response[c].code.substring(0, response[c].code.length - 1)}`], []
-        ).join(',');
-        const script = document.createElement('script');
-        script.textContent = `
-          window.findifyJsonp.push([['extra'], {${content}}]);
-          window.findify.invalidate();
-        `;
-        (document.head || document.documentElement).appendChild(script);
-        script.remove();
-      }
-      // On extension load and on styles saved (onCommit), custom styles are saved and sent.
-      if (type === 'styles') {
-        const { compiled } = response;
-        if (compiled) {
-          console.log('[MJS]: inject new styles');
-          let styleTag = document.createElement('style');
-          document.body.appendChild(styleTag);
-
-          //Remove previous styleTag
-          const element = document.getElementById('findify-styles');
-          if (element && element.parentNode) element.parentNode.removeChild(element);
-
-          styleTag.innerHTML = compiled;
-        }
-      }
-
+      if (type === 'components') injectComponents(response);
+      if (type === 'styles') injectStyles(response);
     });
   }
